214. Set more than 120 lists and arrays to null right after not using any more. This make PA_21_15 CGNF could be solved with only 7G Max Heap 215. Collect Memory Feature added (It is basically a button to restart & reopen the last project user was working on) 216. Collect Memory button also shows memory in use & memory left when moving mouse over it All changes in this block are associated with PrismAlpha1.0.15.jar, Prism Formulation 01.docx, Prism Database 01.docx
Former-commit-id: ff6feb2ab77bfdc694739dfbe8ec39ee97c30b22
Former-commit-id: 268c8c4e55f29102d52ddb849087ffd6812c6362

+++ src/prismRoot/OptionPane_Startup.java
@@ -43,73 +43,111 @@ import prismProject.TextArea_ReadMe;
 public class OptionPane_Startup extends JOptionPane {
 
 	public static void Set_Memory() {
-
 		File jar_file = new File(PrismMain.get_main().getClass().getProtectionDomain().getCodeSource().getLocation().getPath());
 		File restart = new File(FilesHandle.get_temporaryFolder() + "/Restart");	// This folder works as a check. We need the first time restart to activate G1 when we running the jar out of eclipse IDE
 		
-				
-		
+					
 		
 		if (!jar_file.getName().endsWith(".jar")) {	// If running in Eclipse --> always create this bookkeeping folder --> no need to restart the 1st time
 			if (!restart.exists()) restart.mkdirs();
 			restart.deleteOnExit();
 		}
 		
-		
-		
+				
 		
 		File memory_file = new File(FilesHandle.get_temporaryFolder() + "/prism_memory.txt");	// Store the last time MAx Memory is saved by users: just an integer number
-		String delimited = "\t";
 		int previous_max_memory = 0;
+		String previous_project_name = null;
 		try {		
 			List<String> list;
 			list = Files.readAllLines(Paths.get(memory_file.getAbsolutePath()), StandardCharsets.UTF_8);			
 			previous_max_memory = Integer.valueOf(list.get(0));
+			previous_project_name = (list.size() == 2) ? list.get(1) : null;
 		} catch (Exception e) {
 			System.out.println("File prism_memory.txt does not exists");
 			previous_max_memory = 1;
 		}
-		
-		
+				
 		
 		
 		if (!restart.exists()) {
 			restart.mkdirs();
 			try {			
 				// Always restart the 1st time running PRISM to activate G1, also set Max Heap to previously defined max memory (stored in prism_memory.txt) or 1G (in this case we are running the .jar out of eclipse)
-				Memory_File.create_memory_file(memory_file, previous_max_memory);
+				Memory_File.create_memory_file(memory_file, previous_max_memory, previous_project_name);
 				String command_to_execute = "javaw -Xmx" + previous_max_memory + "G -XX:+UseG1GC -jar " + jar_file.getName();
 				Runtime.getRuntime().exec(command_to_execute, null, new File(FilesHandle.get_workingLocation()));
 			} catch (IOException e) {
 			} finally {
 				System.exit(0);
 			}
-		} else {		
-			PrismMain.get_main().setVisible(false);
-			String ExitOption[] = {"Start","Exit"};
-			int response = JOptionPane.showOptionDialog(PrismMain.get_Prism_DesktopPane(), new ScrollPane_Popup(jar_file, memory_file), "Welcome",
-					JOptionPane.DEFAULT_OPTION, JOptionPane.PLAIN_MESSAGE, null, ExitOption, ExitOption[0]);
-			if (response == 0) {
+		} else {			
+			if (previous_project_name == null) {						
+				PrismMain.get_main().setVisible(false);
+				String ExitOption[] = {"Start","Exit"};
+				int response = JOptionPane.showOptionDialog(PrismMain.get_Prism_DesktopPane(), new ScrollPane_Popup(jar_file, memory_file), "Welcome",
+						JOptionPane.DEFAULT_OPTION, JOptionPane.PLAIN_MESSAGE, null, ExitOption, ExitOption[0]);
+				if (response == 0) {
+					restart.delete();
+				} else {
+					restart.delete();
+					System.exit(0);
+				}
+				PrismMain.get_main().setVisible(true);
+			} else {	// When people press the collect memory button in the Panel_Project --> don't need to show the interface to change max memory, open the project instead
 				restart.delete();
-			} else {
-				restart.delete();
-				System.exit(0);
+				PrismMain.get_main().createNewJInternalFrame(previous_project_name);
+				Memory_File.create_memory_file(memory_file, previous_max_memory, null);
 			}
-			PrismMain.get_main().setVisible(true);
 		}	
 	}
+	
+	
+	
+	public static void Restart_Project(String currentProject) {
+		File jar_file = new File(PrismMain.get_main().getClass().getProtectionDomain().getCodeSource().getLocation().getPath());
+		
+		if (jar_file.getName().endsWith(".jar")) {	// If not running in Eclipse --> "Collect Memory" button works
+			File memory_file = new File(FilesHandle.get_temporaryFolder() + "/prism_memory.txt");	// Store the last time MAx Memory is saved by users: just an integer number			
+			int previous_max_memory = 0;
+			try {		
+				List<String> list;
+				list = Files.readAllLines(Paths.get(memory_file.getAbsolutePath()), StandardCharsets.UTF_8);			
+				previous_max_memory = Integer.valueOf(list.get(0));
+			} catch (Exception ex) {
+				System.out.println("File prism_memory.txt does not exists");
+			}
+			
+			if (memory_file.exists()) {
+				memory_file.delete();		// Delete the old file before writing new contents
+			}
+			
+			try (BufferedWriter fileOut = new BufferedWriter(new FileWriter(memory_file))) {			
+				fileOut.write(String.valueOf(previous_max_memory));		
+				fileOut.write( "\n" + currentProject);
+				fileOut.close();
+			} catch (IOException e) {
+				System.err.println(e.getClass().getName() + ": " + e.getMessage());
+			} 		
+			
+			Set_Memory();
+		}
+	}
+	
+	public static String memory_left() {	
+		return formatSize(Runtime.getRuntime().maxMemory() - Runtime.getRuntime().totalMemory() + Runtime.getRuntime().freeMemory());
+	}
+	
+	public static String memory_in_use() {	
+		return formatSize(Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory());
+	}
+
+	private static String formatSize(long v) {
+	    if (v < 1024) return v + " B";
+	    int z = (63 - Long.numberOfLeadingZeros(v)) / 10;
+	    return String.format("%.1f %sB", (double)v / (1L << (z*10)), " KMGTPE".charAt(z));
+	}
 }
-
-
-
-
-
-
-
-
-
-
-
 
 
 
@@ -181,10 +219,10 @@ class ScrollPane_Popup extends JScrollPane {
 	    combo.addActionListener(new AbstractAction() {
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				if (jar_file.getName().endsWith(".jar")) {	// If not running in Eclipse --> do not restart
+				if (jar_file.getName().endsWith(".jar")) {	// If not running in Eclipse --> restart
 					try {
 						String new_max_heap = combo.getSelectedItem().toString();
-						Memory_File.create_memory_file(memory_file, Integer.valueOf(new_max_heap));
+						Memory_File.create_memory_file(memory_file, Integer.valueOf(new_max_heap), null);
 						String command_to_execute = "javaw -Xmx" + new_max_heap + "G -XX:+UseG1GC -jar " + jar_file.getName();
 						Runtime.getRuntime().exec(command_to_execute, null, new File(FilesHandle.get_workingLocation()));
 					} catch (IOException ex) {
@@ -229,25 +267,28 @@ class ScrollPane_Popup extends JScrollPane {
 		setBorder(border);
 		setViewportView(popupPanel);			
 	}
-
-	public static String formatSize(long v) {
+	
+	private static String formatSize(long v) {
 	    if (v < 1024) return v + " B";
 	    int z = (63 - Long.numberOfLeadingZeros(v)) / 10;
 	    return String.format("%.1f %sB", (double)v / (1L << (z*10)), " KMGTPE".charAt(z));
 	}
 	
-	
 }
 
 
+
+
+
 class Memory_File {
-	public static void create_memory_file(File memory_file, int max_heap) {
+	public static void create_memory_file(File memory_file, int max_heap, String previous_project_name) {
 		if (memory_file.exists()) {
 			memory_file.delete();		// Delete the old file before writing new contents
 		}
 		
 		try (BufferedWriter fileOut = new BufferedWriter(new FileWriter(memory_file))) {			
-			fileOut.write(String.valueOf(max_heap));			
+			fileOut.write(String.valueOf(max_heap));		
+			if (previous_project_name != null) fileOut.write( "\n" + previous_project_name);	
 			fileOut.close();
 		} catch (IOException e) {
 			System.err.println(e.getClass().getName() + ": " + e.getMessage());
